requirements.txt
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
html-sanitizer==1.9.3
html5lib==1.1
idna==3.4
lxml==4.9.2
numpy==1.24.2
pandas==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.3.23
requests==2.28.2
six==1.16.0
soupsieve==2.4.1
textile==4.0.2
tzdata==2023.3
urllib3==1.26.15
webencodings==0.5.1
